New archive - load levels from files
I made an update to the level editor and game that makes it easier to add levels. They're now loaded from files. My next step is probably going to make it easier to set the next and previous level warps, I want to set them in the level file, instead of in code.